Integrate IdentityProviderAwareAuthService into LedgerApiServer
