+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/ApiServiceOwner.scala
@@ -9,7 +9,14 @@ import com.daml.api.util.TimeProvider
 import com.daml.buildinfo.BuildInfo
 import com.daml.jwt.JwtTimestampLeeway
 import com.daml.ledger.api.auth.interceptor.AuthorizationInterceptor
-import com.daml.ledger.api.auth.{AuthService, Authorizer}
+import com.daml.ledger.api.auth.{
+  AuthService,
+  Authorizer,
+  IdentityProviderAwareAuthService,
+  IdentityProviderConfigLoader,
+  JwtVerifierLoader,
+}
+import com.daml.ledger.api.domain
 import com.daml.ledger.api.health.HealthChecks
 import com.daml.ledger.configuration.LedgerId
 import com.daml.ledger.participant.state.index.v2.IndexService
@@ -34,7 +41,7 @@ import scalaz.{-\/, \/-}
 
 import java.time.Clock
 import scala.collection.immutable
-import scala.concurrent.ExecutionContextExecutor
+import scala.concurrent.{ExecutionContextExecutor, Future}
 import scala.util.{Failure, Success, Try}
 
 object ApiServiceOwner {
@@ -60,6 +67,7 @@ object ApiServiceOwner {
         _ => None, // Used for Canton rate-limiting,
       ledgerFeatures: LedgerFeatures,
       authService: AuthService,
+      jwtVerifierLoader: JwtVerifierLoader,
       meteringReportKey: MeteringReportKey = CommunityKey,
       jwtTimestampLeeway: Option[JwtTimestampLeeway],
       explicitDisclosureUnsafeEnabled: Boolean = false,
@@ -94,6 +102,13 @@ object ApiServiceOwner {
     )
     // TODO LLP: Consider fusing the index health check with the indexer health check
     val healthChecksWithIndexService = healthChecks + ("index" -> indexService)
+
+    val identityProviderConfigLoader = new IdentityProviderConfigLoader {
+      override def getIdentityProviderConfig(issuer: LedgerId)(implicit
+          loggingContext: LoggingContext
+      ): Future[domain.IdentityProviderConfig] =
+        identityProviderConfigStore.getActiveIdentityProviderByIssuer(issuer)
+    }
 
     for {
       executionSequencerFactory <- new ExecutionSequencerFactoryOwner()
@@ -136,7 +151,11 @@ object ApiServiceOwner {
         config.address,
         config.tls,
         AuthorizationInterceptor(
-          authService,
+          authService = new IdentityProviderAwareAuthService(
+            defaultAuthService = authService,
+            identityProviderConfigLoader = identityProviderConfigLoader,
+            jwtVerifierLoader = jwtVerifierLoader,
+          )(servicesExecutionContext, loggingContext),
           Option.when(config.userManagement.enabled)(userManagementStore),
           servicesExecutionContext,
         ) :: otherInterceptors,
